fix clippy errors
